requirements.txt
Django==5.2.5
djangorestframework==3.16.1
celery==5.5.3
redis==6.4.0
cryptography==45.0.6 